Add webjars integration and update UI components
The code changes consist of integrating webjars into the project, which provides a standard way to manage client-side web libraries. The UI has also been updated with the addition of a navbar and footer. Furthermore, the HomeController has been modified to use WebJarsUtil and the views have been updated to reflect the new design and layout. Import statements and the build.sbt file were also adjusted in this process.