Refactor view paths and improve code organization
The changes primarily involve refactoring the view paths in the SignInController and SignUpController classes. This is part of an effort to improve code organization and readability. Additionally, some import statements in signIn.scala.html and main.scala.html have been modified to optimize code structure and align with the aforementioned path changes.