Implement find and adminUser functions in UserDAOImpl
The 'find' and 'adminUser' methods are now functional and return results from the database. The 'find' method searches for a user by their UUID and the 'adminUser' method searches for a user with the admin role. An extra helper function, 'mapUserFromDBModel', has also been added to map database models to user objects.