Update HomeController and index view
Added AssetsFinder as an implicit parameter in the HomeController and index view. Simplified index view by removing a form and replaced it with an image placeholder. These changes are done in preparation for refining the user interface.

Image is AI generated and simply a placeholder until something can be commissioned.