Add Cookie Use Policy page
A new page detailing the website's Cookie Use Policy has been added to the project. The HomeController is updated to serve the new page, and the footer is adjusted to contain a link to the page. A small adjustment to padding was also made in the main.css file to accommodate this change.