Add terms and conditions view and route
A new terms and conditions view has been introduced with its respective controller action in HomeController. A new route to the terms and conditions page has been added, and it is now accessible through the footer. The terms and conditions HTML file has been created along with updating the footer link.