Revising the model with the assumption that the users of the application will be using the Pan Genome instead of traditional linear assemblies.
Introduced the necessary schema elements for supporting Pan Genome graphs, mapping existing Variants in the haplogroup trees to the canonical pangenome variant and support for indexing and gaining consensus on IBD segments in the network.  Updated the documents to refer to Pan Genome as the target reference.  Also replaced the quality metrics table with something better suited to Pan Genome.